solutions/hello/app/requirements.txt
httpx
fastapi
uvicorn[standard]
pydantic
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-propagator-gcp
python-dotenv